refactor(form): move RegistryContext
ref #276
