+++ packages/dm-core-plugins/src/form/fields/BinaryField.tsx
@@ -1,14 +1,14 @@
-import React from 'react'
-import { useFormContext } from 'react-hook-form'
-import { TStringFieldProps } from '../types'
-import { Button, Label } from '@equinor/eds-core-react'
-import { useRegistryContext } from '../RegistryContext'
 import {
   ErrorResponse,
   TGenericObject,
   useDMSS,
 } from '@development-framework/dm-core'
+import { Button, Label } from '@equinor/eds-core-react'
 import { AxiosError, AxiosResponse } from 'axios'
+import React from 'react'
+import { useFormContext } from 'react-hook-form'
+import { useRegistryContext } from '../context/RegistryContext'
+import { TStringFieldProps } from '../types'
 
 const getTarget = (initialValue: any) => {
   const { idReference } = useRegistryContext()
